revert default securityContext config
HAProxy 2.4 and newer starts as UID 99, and there is nothing we can
safely do by default to fix that when user configures external haproxy.
sysctls config is being reverted because it will not work with
hostNetwork, and can fail as well depending on cluster security
policies.

Added some configuration options and respective drawbacks in the
external haproxy example page.